requirements.txt
flask>=3.0.0
flask-openapi3>=3.1.0
PyJWT>=2.8.0
flask-limiter>=3.5.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
flask-openapi3[swagger]
